Set the required minimum version of PyQt5
If PyQt5 (and Qt5) use DirectShow backend for the media player and the
version is lower than 5.13, it can not use QVideoProbe. As a result,
JS08 will fail to get a steel shot of the incoming video stream with the
following error message:
```
qt.multimedia.plugins.directshow: Unable to connect the video probe!
```
I could confirm the normal execution on Python 3.9 with PyQt 5.15 of the
Anaconda environment. However, according to the documentation, JS08 will
work properly on Python 3.6 with PyQt 5.13 or a higher version.